fix setup
