+++ python/setup_tools/utils/enflame.py
@@ -1,4 +1,5 @@
 import glob
+import importlib.util
 import os
 import shutil
 import sys
@@ -15,6 +16,37 @@ def get_package_data_tools():
         "_triton_gcu300*.so",
         "_triton_gcu400*.so",
     ]
+
+
+def _find_mlir_core_path():
+    """Find MLIR Python bindings (mlir_core) from LLVM env vars or flagtree cache."""
+    search_roots = []
+    for env_var in ("LLVM_SYSPATH", "KURAMA_LLVM_DIR"):
+        val = os.environ.get(env_var, "")
+        if val:
+            search_roots.append(val)
+
+    home = os.environ.get("HOME", os.path.expanduser("~"))
+    flagtree_cache = os.environ.get("FLAGTREE_CACHE_DIR", os.path.join(home, ".flagtree"))
+    enflame_cache = os.path.join(flagtree_cache, "enflame")
+    if os.path.isdir(enflame_cache):
+        for entry in os.listdir(enflame_cache):
+            entry_path = os.path.join(enflame_cache, entry)
+            if os.path.isdir(entry_path) and "llvm" in entry.lower():
+                search_roots.append(entry_path)
+
+    for root in search_roots:
+        candidate = os.path.join(root, "python_packages", "mlir_core")
+        if os.path.isdir(os.path.join(candidate, "mlir")):
+            return candidate
+
+    try:
+        spec = importlib.util.find_spec("mlir")
+        if spec and spec.origin:
+            return None
+    except (ModuleNotFoundError, ValueError):
+        pass
+    return None
 
 
 def install_extension(*args, **kargs):
@@ -71,3 +103,17 @@ def install_extension(*args, **kargs):
             dst_path = dst_dir / src_path.name
             print(f"Copying {src_path} -> {dst_path}")
             shutil.copy2(src_path, dst_path)
+
+    # Copy MLIR Python bindings to build_lib for packaging
+    build_ext = kargs.get('build_ext')
+    if build_ext and hasattr(build_ext, 'build_lib'):
+        mlir_core_path = _find_mlir_core_path()
+        if mlir_core_path:
+            mlir_src = os.path.join(mlir_core_path, "mlir")
+            mlir_dst = os.path.join(build_ext.build_lib, "mlir")
+            if os.path.isdir(mlir_src):
+                if os.path.exists(mlir_dst):
+                    shutil.rmtree(mlir_dst)
+                shutil.copytree(mlir_src, mlir_dst)
+                print(f"Copied MLIR Python bindings from {mlir_src} to {mlir_dst}",
+                      file=sys.stderr)
