+++ setup.py
@@ -2,7 +2,6 @@ import os
 import platform
 import re
 import contextlib
-import importlib.util
 import shlex
 import shutil
 import subprocess
@@ -133,7 +132,7 @@ class BackendInstaller:
         ]
 
 
-# Taken from https://github.com/pytorch/pytorch/blob/master/tools/setup_helpers/env.py
+# Taken from http://minio.idc.sse/public/mirror/github/pytorch/pytorch/blob/master/tools/setup_helpers/env.py
 def check_env_flag(name: str, default: str = "") -> bool:
     return os.getenv(name, default).upper() in ["ON", "1", "YES", "TRUE", "Y"]
 
@@ -191,7 +190,7 @@ class Package:
 
 # json
 def get_json_package_info():
-    url = "https://github.com/nlohmann/json/releases/download/v3.11.3/include.zip"
+    url = "http://minio.idc.sse/public/mirror/github/nlohmann/json/releases/download/v3.11.3/include.zip"
     return Package("json", "", url, "JSON_INCLUDE_DIR", "", "JSON_SYSPATH")
 
 
@@ -257,7 +256,7 @@ def get_llvm_package_info():
     name = f"llvm-{rev}-{system_suffix}"
     # Create a stable symlink that doesn't include revision
     sym_name = f"llvm-{system_suffix}"
-    url = f"https://oaitriton.blob.core.windows.net/public/llvm-builds/{name}.tar.gz"
+    url = f"http://minio.idc.sse/public/mirror/public/llvm-builds/{name}.tar.gz"
     return Package("llvm", name, url, "LLVM_INCLUDE_DIRS", "LLVM_LIBRARY_DIR", "LLVM_SYSPATH", sym_name=sym_name)
 
 
@@ -556,17 +555,6 @@ class CMakeBuild(build_ext):
         subprocess.check_call(["cmake", "--build", "."] + build_args, cwd=cmake_dir)
         subprocess.check_call(["cmake", "--build", ".", "--target", "mlir-doc"], cwd=cmake_dir)
 
-        if helper.flagtree_backend == "enflame":
-            mlir_core_path = _find_mlir_core_path()
-            if mlir_core_path:
-                mlir_src = os.path.join(mlir_core_path, "mlir")
-                mlir_dst = os.path.join(self.build_lib, "mlir")
-                if os.path.isdir(mlir_src):
-                    if os.path.exists(mlir_dst):
-                        shutil.rmtree(mlir_dst)
-                    shutil.copytree(mlir_src, mlir_dst)
-                    print(f"Copied MLIR Python bindings from {mlir_src} to {mlir_dst}", file=sys.stderr)
-
         helper.install_extension(build_ext=self)
 
 
@@ -584,7 +572,7 @@ def download_and_copy_dependencies():
         variable="TRITON_PTXAS_PATH",
         version=NVIDIA_TOOLCHAIN_VERSION["ptxas"],
         url_func=lambda system, arch, version:
-        f"https://developer.download.nvidia.com/compute/cuda/redist/cuda_nvcc/{system}-{arch}/cuda_nvcc-{system}-{arch}-{version}-archive.tar.xz",
+        f"http://minio.idc.sse/public/mirror/nvidia/cuda_nvcc/{system}-{arch}/cuda_nvcc-{system}-{arch}-{version}-archive.tar.xz",
     )
 
     # We download a separate ptxas for blackwell, since there are some bugs when using it for hopper
@@ -595,7 +583,7 @@ def download_and_copy_dependencies():
         variable="TRITON_PTXAS_BLACKWELL_PATH",
         version=NVIDIA_TOOLCHAIN_VERSION["ptxas-blackwell"],
         url_func=lambda system, arch, version:
-        f"https://developer.download.nvidia.com/compute/cuda/redist/cuda_nvcc/{system}-{arch}/cuda_nvcc-{system}-{arch}-{version}-archive.tar.xz",
+        f"http://minio.idc.sse/public/mirror/nvidia/cuda_nvcc/{system}-{arch}/cuda_nvcc-{system}-{arch}-{version}-archive.tar.xz",
     )
     download_and_copy(
         name="cuobjdump",
@@ -605,7 +593,7 @@ def download_and_copy_dependencies():
         variable="TRITON_CUOBJDUMP_PATH",
         version=NVIDIA_TOOLCHAIN_VERSION["cuobjdump"],
         url_func=lambda system, arch, version:
-        f"https://developer.download.nvidia.com/compute/cuda/redist/cuda_cuobjdump/{system}-{arch}/cuda_cuobjdump-{system}-{arch}-{version}-archive.tar.xz",
+        f"http://minio.idc.sse/public/mirror/nvidia/cuda_cuobjdump/{system}-{arch}/cuda_cuobjdump-{system}-{arch}-{version}-archive.tar.xz",
     )
     download_and_copy(
         name="nvdisasm",
@@ -615,7 +603,7 @@ def download_and_copy_dependencies():
         variable="TRITON_NVDISASM_PATH",
         version=NVIDIA_TOOLCHAIN_VERSION["nvdisasm"],
         url_func=lambda system, arch, version:
-        f"https://developer.download.nvidia.com/compute/cuda/redist/cuda_nvdisasm/{system}-{arch}/cuda_nvdisasm-{system}-{arch}-{version}-archive.tar.xz",
+        f"http://minio.idc.sse/public/mirror/nvidia/cuda_nvdisasm/{system}-{arch}/cuda_nvdisasm-{system}-{arch}-{version}-archive.tar.xz",
     )
     crt = "crt" if int(NVIDIA_TOOLCHAIN_VERSION["cudacrt"].split(".")[0]) >= 13 else "nvcc"
     download_and_copy(
@@ -625,7 +613,7 @@ def download_and_copy_dependencies():
         variable="TRITON_CUDACRT_PATH",
         version=NVIDIA_TOOLCHAIN_VERSION["cudacrt"],
         url_func=lambda system, arch, version:
-        f"https://developer.download.nvidia.com/compute/cuda/redist/cuda_{crt}/{system}-{arch}/cuda_{crt}-{system}-{arch}-{version}-archive.tar.xz",
+        f"http://minio.idc.sse/public/mirror/nvidia/cuda_{crt}/{system}-{arch}/cuda_{crt}-{system}-{arch}-{version}-archive.tar.xz",
     )
     download_and_copy(
         name="cudart",
@@ -634,7 +622,7 @@ def download_and_copy_dependencies():
         variable="TRITON_CUDART_PATH",
         version=NVIDIA_TOOLCHAIN_VERSION["cudart"],
         url_func=lambda system, arch, version:
-        f"https://developer.download.nvidia.com/compute/cuda/redist/cuda_cudart/{system}-{arch}/cuda_cudart-{system}-{arch}-{version}-archive.tar.xz",
+        f"http://minio.idc.sse/public/mirror/nvidia/cuda_cudart/{system}-{arch}/cuda_cudart-{system}-{arch}-{version}-archive.tar.xz",
     )
     download_and_copy(
         name="cupti",
@@ -643,7 +631,7 @@ def download_and_copy_dependencies():
         variable="TRITON_CUPTI_INCLUDE_PATH",
         version=NVIDIA_TOOLCHAIN_VERSION["cupti"],
         url_func=lambda system, arch, version:
-        f"https://developer.download.nvidia.com/compute/cuda/redist/cuda_cupti/{system}-{arch}/cuda_cupti-{system}-{arch}-{version}-archive.tar.xz",
+        f"http://minio.idc.sse/public/mirror/nvidia/cuda_cupti/{system}-{arch}/cuda_cupti-{system}-{arch}-{version}-archive.tar.xz",
     )
     download_and_copy(
         name="cupti",
@@ -652,7 +640,7 @@ def download_and_copy_dependencies():
         variable="TRITON_CUPTI_LIB_PATH",
         version=NVIDIA_TOOLCHAIN_VERSION["cupti"],
         url_func=lambda system, arch, version:
-        f"https://developer.download.nvidia.com/compute/cuda/redist/cuda_cupti/{system}-{arch}/cuda_cupti-{system}-{arch}-{version}-archive.tar.xz",
+        f"http://minio.idc.sse/public/mirror/nvidia/cuda_cupti/{system}-{arch}/cuda_cupti-{system}-{arch}-{version}-archive.tar.xz",
     )
 
 
@@ -669,37 +657,6 @@ else:
     backends = [*BackendInstaller.copy(["nvidia", "amd"]), *BackendInstaller.copy_externals()]
 
 #backends = [*BackendInstaller.copy(["nvidia", "amd"]), *BackendInstaller.copy_externals()]
-
-
-def _find_mlir_core_path():
-    """Find MLIR Python bindings (mlir_core) from LLVM env vars or flagtree cache."""
-    search_roots = []
-    for env_var in ("LLVM_SYSPATH", "KURAMA_LLVM_DIR"):
-        val = os.environ.get(env_var, "")
-        if val:
-            search_roots.append(val)
-
-    home = os.environ.get("HOME", os.path.expanduser("~"))
-    flagtree_cache = os.environ.get("FLAGTREE_CACHE_DIR", os.path.join(home, ".flagtree"))
-    enflame_cache = os.path.join(flagtree_cache, "enflame")
-    if os.path.isdir(enflame_cache):
-        for entry in os.listdir(enflame_cache):
-            entry_path = os.path.join(enflame_cache, entry)
-            if os.path.isdir(entry_path) and "llvm" in entry.lower():
-                search_roots.append(entry_path)
-
-    for root in search_roots:
-        candidate = os.path.join(root, "python_packages", "mlir_core")
-        if os.path.isdir(os.path.join(candidate, "mlir")):
-            return candidate
-
-    try:
-        spec = importlib.util.find_spec("mlir")
-        if spec and spec.origin:
-            return None
-    except (ModuleNotFoundError, ValueError):
-        pass
-    return None
 
 
 def get_package_dirs():
@@ -956,7 +913,7 @@ setup(
     zip_safe=False,
     # for PyPI
     keywords=["Compiler", "Deep Learning"],
-    url="https://github.com/flagos-ai/FlagTree/",
+    url="http://minio.idc.sse/public/mirror/github/flagos-ai/FlagTree/",
     python_requires=PYTHON_REQUIRES,
     classifiers=CLASSIFIERS,
     test_suite="tests",
